requirements.txt
pyserial
adafruit-io